+++ com.unity.postprocessing/PostProcessing/Editor/EffectListEditor.cs
@@ -5,6 +5,8 @@ using UnityEngine;
 using UnityEngine.Assertions;
 using UnityEngine.Rendering.PostProcessing;
 
+#pragma warning disable CS0618 // Type or member is obsolete
+
 namespace UnityEditor.Rendering.PostProcessing
 {
     /// <summary>
@@ -343,3 +345,5 @@ namespace UnityEditor.Rendering.PostProcessing
         }
     }
 }
+
+#pragma warning restore CS0618 // Type or member is obsolete