+++ com.unity.postprocessing/PostProcessing/Editor/PostProcessProfileEditor.cs
@@ -1,5 +1,7 @@
 using UnityEngine.Rendering.PostProcessing;
 
+#pragma warning disable CS0618 // Type or member is obsolete
+
 namespace UnityEditor.Rendering.PostProcessing
 {
     [CustomEditor(typeof(PostProcessProfile))]
@@ -27,3 +29,5 @@ namespace UnityEditor.Rendering.PostProcessing
         }
     }
 }
+
+#pragma warning restore CS0618 // Type or member is obsolete