+++ com.unity.postprocessing/PostProcessing/Editor/Tools/VolumeFactory.cs
@@ -1,6 +1,8 @@
 using UnityEngine;
 using UnityEngine.Rendering.PostProcessing;
 
+#pragma warning disable CS0618 // Type or member is obsolete
+
 namespace UnityEditor.Rendering.PostProcessing
 {
     internal static class VolumeFactory
@@ -19,3 +21,5 @@ namespace UnityEditor.Rendering.PostProcessing
         }
     }
 }
+
+#pragma warning restore CS0618 // Type or member is obsolete