+++ com.unity.postprocessing/PostProcessing/Runtime/PostProcessDebug.cs
@@ -1,3 +1,5 @@
+#pragma warning disable CS0618 // Type or member is obsolete
+
 namespace UnityEngine.Rendering.PostProcessing
 {
     /// <summary>
@@ -158,3 +160,5 @@ namespace UnityEngine.Rendering.PostProcessing
         }
     }
 }
+
+#pragma warning restore CS0618 // Type or member is obsolete